Added motor brackets for the GRITSBot's motor board
These motor brackets are made out of PCB material like the main
and motor board of the robot. Each bracket is glued to the motor
board and functions as caster pins so that the robot is statically
stable. Note that these motor brackets can be replaced with the
3D printed part that also functions as charging coil mount (see
folder '3d_model' in this repository.