Fixed RGB LED setup on main board.
The RGB LEDs were connected to GPIO16, which on some versions
of the ESP8266 is connected to the RESET pin and can't be used
as general purpose pin after all. The RGB LEDs are now connected
to GPIO5 - the same pin as the status LED. Both types of LEDs can
be used together on the same pin.